gesture-detection/requirements.txt
absl-py==2.2.2
attrs==25.3.0
bidict==0.23.1
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.57.0
h11==0.14.0
jax==0.6.0
jaxlib==0.6.0
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.6
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-osc==1.9.3
python-socketio==5.13.0
pytweening==1.2.0
rubicon-objc==0.5.0
scipy==1.15.2
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.1
wsproto==1.2.0
